smart-hire-ai/requirements.txt
streamlit
huggingface_hub
pillow
spacy
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
